[modbus] connection closing behaviour finetuned
The binding closes TCP/serial connections connections as per user
configuration. It is either
- every time, immediately after a transaction (reconnectAfterMillis=0)
- *after* a read/write transaction, if the connection has been open "too
  long" (configurable with reconnectAfterMillis)

We have an obvious downside to this simple logic -- the connection
can remain open "indefinitely" if there are no transactions occurring.
With Modbus we are quite often dealing with PLCs and other embedded
systems with limited resources, and "reserving" the connection is not
always something we want.

Previously (2.5.x branch) connections were also closed when
- a regular poll task was unregistered (i.e. we stop reading regularly
from a modbus server ("slave"). Since most users have regular polling in
place, so this ensured that connections do get closed.

In this commit, the behaviour is adjusted such that connections are
closed when last communication interface pointing to the server/serial
port (i.e. "endpoint") is close()'d.

With modbus binding this basically means when the tcp or serial thing
is removed / disabled, the connections is closed/freed, but only if it
is the last thing pointing to that server or serial port.

Since modbus.sunspec binding reuses modbus serial & tcp endpoint things,
the same note applies for modbus.sunspec binding.

This is change in functional behaviour but in a way is logical. We can
further introduce to have "delayed"/"deferred"/"debounce" connection
closing connections as per the setting reconnectAfterMillis even in
situations where communication interface is still open.

Signed-off-by: Sami Salonen <ssalonen@gmail.com>